janky

+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import Head from "next/head";
 import { Crossword } from "@/components/crossword";
 import React, { useCallback, useEffect, useState } from "react";
 import useApi from "@/hooks/useApi";
-import { Attempt, Numcross, Scratch } from "@/types/types";
+import { Attempt, Numcross, Scratch, Solve } from "@/types/types";
 import { isAttemptFull } from "@/utils";
 import { toast } from "react-hot-toast";
 import useModal from "@/hooks/useModal";
@@ -10,14 +10,17 @@ import { Numpad } from "@/components/numpad";
 import { isEqual } from "lodash";
 import SolvedOverlay from "@/common/solved_overlay";
 import { useUser } from "@supabase/auth-helpers-react";
+import { mineAttempt, mineSolve, storeSolve } from "@/hooks/useStorage";
 
 export default function Home() {
-  const { getTodaysNumcross, checkAttempt, updateAttempt } = useApi();
+  const { getTodaysNumcross, checkAttempt, logSolve, updateAttempt } = useApi();
   const [numcross, setNumcross] = useState<Numcross | null>(null);
   const [scratch, setScratch] = useState<Scratch>({});
   const [attempt, setAttempt] = useState<Attempt | null>(null);
   const [pageError, setPageError] = useState<string | null>(null);
+  const [shouldPopOff, setShouldPopOff] = useState(true);
   const [hasSolved, setHasSolved] = useState<boolean>(false);
+  const [solve, setSolve] = useState<Solve | null>(null);
   const [SolvedModal, openSolved, closeSolved] = useModal();
   const [lastAttempt, setLastAttempt] = useState<Attempt | null>(null);
   const user = useUser();
@@ -25,22 +28,51 @@ export default function Home() {
   useEffect(() => {
     // Helper function to do the get
     const performGet: () => Promise<void> = async () => {
-      localStorage.clear();
+      console.log(
+        !!getTodaysNumcross,
+        user?.id,
+        !!logSolve,
+        !!updateAttempt,
+        shouldPopOff
+      );
       const result = await getTodaysNumcross(user?.id);
       if (!result) {
         setPageError("Error loading today's numcross");
         return;
       }
-      const { numcross, attempt: attemptFetch } = result;
+      // This is to handle a user who did the puzzle logged out and then
+      // made an account
+      const { numcross, attempt: attemptFetch, solve: existSolve } = result;
       setNumcross(numcross);
-      setAttempt(
-        attemptFetch || {
-          puzzleId: numcross.id,
-          scratch: {},
-          startTime: new Date().toISOString(),
-          hasCheated: false,
-        }
-      );
+      const preattempt = mineAttempt(numcross.id);
+      const presolved = mineSolve(numcross.id);
+      if (shouldPopOff && (!!existSolve || (preattempt && presolved))) {
+        setShouldPopOff(false);
+      }
+      if (presolved && preattempt && user?.id) {
+        setSolve(presolved);
+        await updateAttempt(preattempt, user.id);
+        await logSolve(presolved, user.id);
+        setAttempt(
+          attemptFetch || {
+            puzzleId: numcross.id,
+            scratch: preattempt.scratch,
+            startTime: presolved.startTime,
+            hasCheated: presolved.didCheat,
+          }
+        );
+        // TODO: More granular way of doing this
+        localStorage.clear();
+      } else {
+        setAttempt(
+          attemptFetch || {
+            puzzleId: numcross.id,
+            scratch: {},
+            startTime: new Date().toISOString(),
+            hasCheated: false,
+          }
+        );
+      }
       // Explicitly update the scratch, note that usually data
       // flows in the reverse direction
       if (attemptFetch) {
@@ -49,7 +81,7 @@ export default function Home() {
     };
 
     performGet();
-  }, [getTodaysNumcross, user?.id]);
+  }, [getTodaysNumcross, user?.id, logSolve, updateAttempt, shouldPopOff]);
 
   // Effect to make sure the "scratch" is updated in the attempt
   useEffect(() => {
@@ -65,13 +97,21 @@ export default function Home() {
   const doCheckAttempt = useCallback(async () => {
     if (!attempt || hasSolved) return;
     const apiResult = await checkAttempt(attempt, user?.id);
+    if (apiResult?.solve) {
+      if (!solve) {
+        setSolve(apiResult.solve);
+        if (!user?.id) {
+          storeSolve(apiResult.solve);
+        }
+      }
+    }
     if (apiResult?.correct) {
       setHasSolved(true);
     } else {
       toast("Puzzle incorrect", { icon: "😑" });
     }
     return;
-  }, [attempt, checkAttempt, hasSolved, setHasSolved, user?.id]);
+  }, [attempt, checkAttempt, hasSolved, setHasSolved, user?.id, solve]);
 
   // Update the attempt over API on change
   // Check the attempt if it's full
@@ -128,7 +168,7 @@ export default function Home() {
         />
         <Numpad editable={false} />
         <SolvedModal>
-          <SolvedOverlay closeModal={closeSolved} />
+          <SolvedOverlay closeModal={closeSolved} solve={solve} />
         </SolvedModal>
       </div>
     </>
